+++ src/pages/Index.tsx
@@ -1,8 +1,5 @@
 import Header from '@/components/Header';
 import Hero from '@/components/Hero';
-import About from '@/components/About';
-import Portfolio from '@/components/Portfolio';
-import Contact from '@/components/Contact';
 import Footer from '@/components/Footer';
 import MetaTags from '@/components/SEO/MetaTags';
 import StructuredData from '@/components/SEO/StructuredData';
@@ -21,9 +18,6 @@ const Index = () => {
         <Header />
         <main>
           <Hero />
-          <About />
-          <Portfolio />
-          <Contact />
         </main>
         <Footer />
       </div>
